requirements.txt
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.1
faiss-cpu>=1.7.2  # or faiss-gpu for GPU support
matplotlib>=3.4.0
scikit-learn>=0.24.2
pandas>=1.3.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pytest>=6.2.5
black>=21.7b0
isort>=5.9.3
mypy>=0.910
tqdm>=4.65.0